Timeline working, commiting in case of due to Ubuntu presenting some issues.
